Remove abstract application class layer and move documentation.
