+++ docs/developer/index.md
@@ -14,4 +14,5 @@ getting-started
 coding-conventions
 dependency-management
 benchmark
+async_programming
 ```
