+++ src/beamlime/__main__.py
@@ -5,13 +5,13 @@
 def main() -> None:
     from beamlime.executables.options import build_arg_parser
     from beamlime.executables.prototypes import (
+        default_prototype_factory,
         event_generator_arg_parser,
-        mini_prototype_factory,
         run_standalone_prototype,
         visualization_arg_parser,
     )
 
-    factory = mini_prototype_factory()
+    factory = default_prototype_factory()
 
     arg_parser = build_arg_parser()
     event_generator_arg_parser(arg_parser)
