+++ src/beamlime/applications/base.py
@@ -1,9 +1,20 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2024 Scipp contributors (https://github.com/scipp)
 import asyncio
-from abc import ABC, abstractmethod, abstractproperty
+from abc import ABC, abstractmethod
+from contextlib import contextmanager
 from dataclasses import dataclass
-from typing import Any, Callable, List, Optional
+from typing import (
+    Any,
+    Awaitable,
+    Callable,
+    Coroutine,
+    Generator,
+    List,
+    Optional,
+    Protocol,
+    runtime_checkable,
+)
 
 from ..logging import BeamlimeLogger
 from ..logging.mixins import LogMixin
@@ -68,15 +79,6 @@ class DaemonInterface(LogMixin, ABC):
         ...
 
 
-@dataclass
-class BeamlimeMessage:
-    """A message object that can be exchanged between daemons or handlers."""
-
-    content: Any
-    sender: type = Any
-    receiver: type = Any
-
-
 class HandlerInterface(LogMixin, ABC):
     """Base class for message handlers.
 
@@ -88,8 +90,146 @@ class HandlerInterface(LogMixin, ABC):
     logger: BeamlimeLogger
 
 
-class ApplicationInterface(LogMixin, ABC):
-    """Base class for applications.
+@runtime_checkable
+class MessageProtocol(Protocol):
+    content: Any
+    sender: type
+    receiver: type
+
+
+class MessageRouter(DaemonInterface):
+    """A message router that routes messages to handlers."""
+
+    @dataclass
+    class StopRouting:
+        """A message to break the routing loop."""
+
+        content: Optional[Any]
+        sender: type
+        receiver: type
+
+    def __init__(self):
+        from queue import Queue
+
+        self.handlers: dict[type, List[Callable[[MessageProtocol], Any]]] = dict()
+        self.awaitable_handlers: dict[
+            type, List[Callable[[MessageProtocol], Awaitable[Any]]]
+        ] = dict()
+        self.message_pipe = Queue()
+        self._break_routing_loop = False  # Break the routing loop flag
+
+    def break_routing_loop(self, message: Optional[MessageProtocol] = None) -> None:
+        if not isinstance(message, self.StopRouting):
+            raise TypeError(
+                f"Expected message of type {self.StopRouting}, got {type(message)}."
+            )
+        self._break_routing_loop = True
+
+    @contextmanager
+    def _handler_wrapper(
+        self, handler: Callable[..., Any], message: MessageProtocol
+    ) -> Generator[None, None, None]:
+        import warnings
+
+        try:
+            self.debug(f"Routing event {type(message)} to handler {handler}...")
+            yield
+        except Exception as e:
+            warnings.warn(f"Failed to handle event {type(message)}", stacklevel=2)
+            raise e
+        else:
+            self.debug(f"Routing event {type(message)} to handler {handler} done.")
+
+    def _register(
+        self,
+        *,
+        handler_list: dict[
+            type[MessageProtocol], List[Callable[[MessageProtocol], Any]]
+        ],
+        event_tp: type[MessageProtocol],
+        handler: Callable[[MessageProtocol], Any],
+    ):
+        if event_tp in handler_list:
+            handler_list[event_tp].append(handler)
+        else:
+            handler_list[event_tp] = [handler]
+
+    def register_awaitable_handler(
+        self, event_tp, handler: Callable[[MessageProtocol], Coroutine[Any, Any, Any]]
+    ):
+        self._register(
+            handler_list=self.awaitable_handlers, event_tp=event_tp, handler=handler
+        )
+
+    def register_handler(self, event_tp, handler: Callable[[MessageProtocol], Any]):
+        self._register(handler_list=self.handlers, event_tp=event_tp, handler=handler)
+
+    def _collect_results(self, result: Any) -> List[MessageProtocol]:
+        """Append or extend ``result`` to ``self.message_pipe``.
+
+        It filters out non-BeamlimeMessage objects from ``result``.
+        """
+        if isinstance(result, MessageProtocol):
+            return [result]
+        elif isinstance(result, tuple):
+            return list(_msg for _msg in result if isinstance(_msg, MessageProtocol))
+        else:
+            return []
+
+    async def route(self, message: MessageProtocol) -> None:
+        # Synchronous handlers
+        results = []
+        for handler in (handlers := self.handlers.get(type(message), [])):
+            await asyncio.sleep(0)  # Let others use the event loop.
+            with self._handler_wrapper(handler, message):
+                results.extend(self._collect_results(handler(message)))
+
+        # Asynchronous handlers
+        for handler in (
+            awaitable_handlers := self.awaitable_handlers.get(type(message), [])
+        ):
+            with self._handler_wrapper(handler, message):
+                results.extend(self._collect_results(await handler(message)))
+
+        # No handlers
+        if not (handlers or awaitable_handlers):
+            import warnings
+
+            warnings.warn(
+                f"No handler for event {type(message)}. Ignoring...", stacklevel=2
+            )
+
+        # Re-route the results
+        for result in results:
+            self.message_pipe.put(result)
+
+    async def run(self) -> None:
+        """Message router daemon."""
+        data_monitor = self.data_pipe_monitor(
+            self.message_pipe,
+            timeout=0.1,
+            interval=0.1,
+            size_monitor_func=self.message_pipe.qsize,
+        )
+
+        while not self._break_routing_loop:
+            if await data_monitor():
+                message = self.message_pipe.get()
+                await self.route(message)
+                await asyncio.sleep(0)
+
+        self.debug("Breaking routing loop. Routing the rest of the message...")
+        while not self.message_pipe.empty():
+            await self.route(self.message_pipe.get())
+        self.debug("Routing the rest of the message done.")
+
+    async def send_message_async(self, message: MessageProtocol) -> None:
+        self.message_pipe.put(message)
+        await asyncio.sleep(0)
+
+
+class Application(LogMixin):
+    """Application class.
 
     Main Responsibilities:
         - Create/retrieve event loop.
@@ -99,24 +239,36 @@ class ApplicationInterface(LogMixin, ABC):
 
     """
 
-    logger: BeamlimeLogger
-
-    def __init__(self) -> None:
+    def __init__(self, logger: BeamlimeLogger, message_router: MessageRouter) -> None:
         import asyncio
 
         self.loop: asyncio.AbstractEventLoop
         self.tasks: List[asyncio.Task]
+        self.logger = logger
+        self.message_router = message_router
+        self.daemons: List[DaemonInterface] = [self.message_router]
+        self.register_handling_method(
+            self.message_router.StopRouting, self.message_router.break_routing_loop
+        )
         super().__init__()
 
-    @abstractmethod
-    def register_handling_methods(self):
-        """Register handlers to the application."""
-        ...
+    def register_handling_method(
+        self, event_tp: type[MessageProtocol], handler: Callable[[MessageProtocol], Any]
+    ) -> None:
+        """Register handlers to the application message router."""
+        if asyncio.iscoroutinefunction(handler):
+            self.message_router.register_awaitable_handler(event_tp, handler)
 
-    @abstractproperty
-    def daemons(self) -> tuple[DaemonInterface, ...]:
-        """Collection of daemons that are running in the application."""
-        ...
+        self.message_router.register_handler(event_tp, handler)
+
+    def register_daemon(self, daemon: DaemonInterface) -> None:
+        """Register a daemon to the application.
+
+        Registered daemons will be scheduled in the event loop
+        as :func:`~Application.run` method is called.
+        The future of the daemon will be collected in the ``self.tasks`` list.
+        """
+        self.daemons.append(daemon)
 
     def run(self):
         """
@@ -125,56 +277,8 @@ class ApplicationInterface(LogMixin, ABC):
         It retrieves or creates an event loop
         and schedules all coroutines(run methods) of its daemons.
 
-        Notes
-        -----
-        **Debugging log while running async daemons under various circumstances.**
-
-        :func:`asyncio.get_event_loop` vs :func:`asyncio.new_event_loop`
-
-        1. :func:`asyncio.get_event_loop`
-            ``get_event_loop`` will always return the current event loop.
-            If there is no event loop set in the thread, it will create a new one
-            and set it as a current event loop of the thread, and return the loop.
-            Many of ``asyncio`` free functions internally use ``get_event_loop``,
-            i.e. :func:`asyncio.create_task`.
-
-            :strong:`Things to consider while using` :func:`asyncio.get_event_loop`.
-
-            * ``asyncio.create_task`` does not guarantee
-                whether the current loop is/will be alive until the task is done.
-                You may use ``run_until_complete`` to make sure the loop is not closed
-                until the task is finished.
-                When you need to throw multiple async calls to the loop,
-                use ``asyncio.gather`` to merge all the tasks like in this method.
-            * ``close`` or ``stop`` might accidentally destroy/interrupt
-                other tasks running in the same event loop.
-                i.e. You can accidentally destroy the event loop of a jupyter kernel.
-            * *1* :class:`RuntimeError` if there has been an event loop set in the
-                thread object before but it is now removed.
-
-        2. :func:`asyncio.new_event_loop`
-            ``asyncio.new_event_loop`` will always return the new event loop,
-            but it is not set it as a current loop of the thread automatically.
-
-            However, sometimes it is automatically handled within the thread,
-            and it caused errors which was hard to debug under ``pytest`` session.
-            For example,
-
-            * The new event loop was not closed properly as it is destroyed.
-            * The new event loop was never started until it is destroyed.
-
-
-            ``Traceback`` of ``pytest`` did not show
-            where exactly the error is from in those cases.
-            It was resolved by using :func:`asyncio.get_event_loop`,
-            or manually closing the event loop at the end of the test.
-
-            :strong:`When to use` :func:`asyncio.new_event_loop`.
-
-            * :func:`asyncio.get_event_loop` raises :class:`RuntimeError` *1*.
-            * Multi-threads.
-
-        Please note that the loop object might need to be **closed** manually.
+        See :doc:`/developer/async_programming` for more details about
+        why it handles the event loop like this.
 
         """
         import asyncio
@@ -182,7 +286,6 @@ class ApplicationInterface(LogMixin, ABC):
 
         from beamlime.core.schedulers import temporary_event_loop
 
-        self.register_handling_methods()
         self.info('Start running %s...', self.__class__.__qualname__)
         start = time.time()
         with temporary_event_loop() as loop:
@@ -191,4 +294,5 @@ class ApplicationInterface(LogMixin, ABC):
             self.tasks = [loop.create_task(coro) for coro in daemon_coroutines]
             if not loop.is_running():
                 loop.run_until_complete(asyncio.gather(*self.tasks))
+
         self.info('Finished running %s...', time.time() - start)
